+++ CHANGELOG.md
@@ -12,10 +12,12 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Added
 
 - Ability to print config for a file. [#158]
+- `no-css-comments` linter rule. [#235]
 - Debug mode with comprehensive logging. [#199]
 - Support for `package.json` configuration file. [#245]
 
 [#158]: https://github.com/AdguardTeam/AGLint/issues/158
+[#235]: https://github.com/AdguardTeam/AGLint/issues/235
 [#199]: https://github.com/AdguardTeam/AGLint/issues/199
 [#245]: https://github.com/AdguardTeam/AGLint/issues/245